file structure/namespace cleanup
