+++ PromienVK/src/asset/format/model.cpp
@@ -1,0 +1,9 @@
+#include "model.hpp"
+
+namespace asset {
+	namespace format {
+		Model::Model() {}
+
+		Model::~Model() {}
+	}
+}
