--- PromienVK/src/asset/model.cpp
@@ -1,7 +1,0 @@
-#include "model.hpp"
-
-namespace asset {
-	Model::Model(){}
-
-	Model::~Model(){}
-}
