+++ PromienVK/src/core/pipeline/boilerplate.cpp
