+++ PromienVK/src/core/pipeline/boilerplate.hpp
